Extract perforation and stencil in their own modules rename mask.py module
